\name{BRCA}
\alias{BRCA}


\docType{data}
\title{BRCA dataset for MLPA analysis }

\usage{data(BRCA)}

\description{
This data set is used in the vignette MLPAstats. You can access it in the gui 
under the ''File->load demo´´ menu.
}

\format{
  BRCAcontrols contains data for control subjects
  BRCAcases data for case subjects
  reference.probes encodes the numberig of the probes to be taken as internal control for normalization  
  size provides size of the probes
}

\examples{
data(BRCA)
}
\keyword{datasets}
