\name{gui.mlpa}
\alias{gui.mlpa}

\title{GUI for MLPAstats }
\description{
Launches the an interface for the analysis of MLPA data.
}
\usage{
gui.mlpa()
}

\details{
see vignette("MLPAstats") for a set-by-etep guide
}

\references{
Caceres, A, et al, MLPAstats: An R GUI package for the integrated analysis of copy number alterations using MLPA data,
BMC Bioinformatics, under review JAN-2010

Gonzalez, J.R., et al, Probe-specific mixed-model approach
to detect copy number differences using multiplex ligation-dependent
probe amplification. BMC Bioinformatics, 9, 261. 2008. }
    
\examples{
\dontrun{
## Start the GUI
gui.mlpa()
}

}

\keyword{gui}
