\name{MLPAstats-Ms}
\alias{Ms}
\alias{getInfo}
\alias{getProbes}
\alias{getPeaks}
\alias{getSize}
\alias{getProbesControl}
\alias{getResults}
\alias{getConfig}
\alias{getNormalize}

\title{MLPAstats functions for GUI interface}

\description{Functions to get information about the object 'Ms' when GUI interface is used}

\usage{
getInfo()
getProbes()
getPeaks(name.probe, individuals="controls")
getSize()
getProbesControl()
getConfig()
getResults()
getNormalize()
}

\arguments{
  \item{name.probe}{vector with names of probes}
  \item{individuals}{whether individuals should be takes as "controls" or "tests"}
}

\details{These are called when GUI interface is used}

\keyword{utilities}
