\name{MLPAstats-package}
\alias{MLPAstats-package}
\alias{MLPAstats}
\docType{package}
\title{
MLPAstats 
}
\description{
 MLPAstats is a software designed to test copy number differences between case and control samples, identified with multiplex-dependent probe amplification (MLPA) data.
}
\details{
\tabular{ll}{
Package: \tab MLPAstats\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-28-01\cr
License: \tab GPL version 2 or newer\cr
LazyLoad: \tab yes\cr
}
The package is implemented in a GUI that can be called with the function gui.mlpa(). 
Further information can be obtained from the vignette: "MLPAstats.pdf" 
}
\author{
Alejandro Caceres and Juan R. Gonzalez

Maintainer: Alejandro Caceres<acaceres@creal.cat> and Juan R Gonzalez <jrgonzalez@creal.cat>
}
\references{

Caceres, A, et al, MLPAstats: An R GUI package for the integrated analysis of copy number alterations using MLPA data,
BMC Bioinformatics, under review JAN-2010

Gonzalez, J.R., et al, Probe-specific mixed-model approach
to detect copy number differences using multiplex ligation-dependent
probe amplification. BMC Bioinformatics, 9, 261. 2008.

}

\examples{
gui.mlpa()
}

\keyword{ package }
