\name{MLPAvalidation}
\alias{MLPAvalidation}
\alias{controls}
\alias{tests}
\alias{size}
\alias{probes.control}

\docType{data}
\title{MLPA dataset for validation study}

\usage{data(MLPAvalidation)}

\description{
  This data set is used in the vignette ''MLPAstats´´ to illustrate the use of 
  software from the comand line. 
}

\format{
  controls contains data for control subjects
  tests data for test subjects
  probes.control numberig of the probes to be taken as internal control for normalization  
  size provides size of the probes
}

\examples{
data(MLPAvalidation)
}
\keyword{datasets}
